refactor(pipeline): extract phase functions into pipeline/phases/
Split the per-phase implementations out of the 1307-line orchestrator.py
into a phases/ package — ingestion (_run_ingestion + the ProcessPool
worker + _timed_step), git (_run_git_indexing), analysis (dead-code +
health + decision extraction), and generation (run_generation) — plus a
shared phases/_common.py for the _phase_done helper.

orchestrator.py keeps run_pipeline() and PipelineResult, and re-imports
run_generation and the private _run_health_analysis so both stay
importable from repowise.core.pipeline.orchestrator (the package
__init__ re-exports the former; cli/commands/upgrade_flow.py imports the
latter directly).

Fixed the two N806 uppercase locals and the one SIM105 try/except/pass
that were carried along in the moved code, so the new modules are
lint-clean.

No behavior change. Ruff clean, no new mypy errors; pipeline unit +
integration + reasoning tests green (21 passed).
