+++ packages/core/src/repowise/core/pipeline/phases/__init__.py
@@ -1,0 +1,1 @@
+"""Pipeline phase implementations (ingestion, git, analysis, generation)."""
